Move network switch to parent component
